new message:
Info.serverURL has not been set. The url https://asdf.adsf.adsf is in a
domain that requires Cross Origin Resource Sharing (CORS) access from
this page, and that domain is not recognized by SwingJS as allowing CORS
access. If the server does allow CORS, then the developer of this page
must make sure that J2S.addDirectDatabaseCall(path) is called for that
domain -- for example, J2S.addDirectDatabaseCall('stolaf.edu'). And if
that domain does not allow CORS, then the developer should set
Info.serverURL appropriately or contact the administrator of that domain
to see if CORS can be allowed.