Fix for splash JFrame creation before completion of Tracker initialization
If a new thread is started during the first JFrame initialization, then
it is possible that the new thread will be created before the thread
group has gotten its reference to the JAppletViewer, which it needs to
manage the overall set of frames created by an application. Solution was
to steal the applet viewer from ThreadGroup.秘html5Applet, which is
assigned first.