DataTable fix, DatasetManager option
In DataTableColumnModel there is a method getColumn(int) that overrides
DefaultTableColumnModel.getColumn(int) by adding the capability to
create a new TableColumn if the requested index is out of range and also
process the DataTableColumn.isSizeSet flag. This is fine for rendering
(is it?) but in general, we don't want to be hacking the columnIndex
this way or producing the setting of the flag before it is necessary. So
getTableColumn(i) now calls super.getColumn(i).

DatasetManager could send information about column ordering, but it does
not. I sketched in a partial solution, but it is just commented out for
now.

in DataToolTable.getWorkingData there is a call to getColumnName() that
uselessly creates a new column just because it is calling
getColumnName() with a 1 prematurely, before the x-data has been loaded
into the dataset.