bug #143
Double-clicking a Tracker node in the Library Browser doesn't load the
file in JS. It loads fine when you click the Open button. Works in Java.

I cannot remember why Double-click does not always work, I think because
you have to hit it just right on the label, so I set it to open if you
click once on the icon. But there was a bug there.